Optimize React Rendering
