+++ React/src/constants.ts
@@ -1,0 +1,1 @@
+export const backendURL = "http://localhost:5020/";
